dra ut dto-er i egne filer fiks tittel i swagger
