+++ src/main/kotlin/no/nav/familie/ks/barnehagelister/rest/BarnehagelisterController.kt
@@ -8,6 +8,7 @@ import io.swagger.v3.oas.annotations.responses.ApiResponses
 import jakarta.servlet.http.HttpServletRequest
 import jakarta.validation.Valid
 import no.nav.familie.ks.barnehagelister.kontrakt.FormV1
+import no.nav.familie.ks.barnehagelister.kontrakt.KindergartenlistResponse
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.http.ResponseEntity
 import org.springframework.validation.BindingResult
@@ -16,7 +17,6 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
-import java.time.LocalDateTime
 import java.util.UUID
 
 @ProtectedWithClaims(
@@ -34,7 +34,7 @@ interface BarnehagelisterController {
                 content = [
                     Content(
                         mediaType = "application/json",
-                        schema = Schema(implementation = BarnehagelisteResponse::class),
+                        schema = Schema(implementation = KindergartenlistResponse::class),
                     ),
                 ],
             ),
@@ -44,7 +44,7 @@ interface BarnehagelisterController {
                 content = [
                     Content(
                         mediaType = "application/json",
-                        schema = Schema(implementation = BarnehagelisteResponse::class),
+                        schema = Schema(implementation = KindergartenlistResponse::class),
                     ),
                 ],
             ),
@@ -89,7 +89,7 @@ interface BarnehagelisterController {
         @Valid @RequestBody formV1: FormV1,
         bindingResult: BindingResult,
         request: HttpServletRequest,
-    ): ResponseEntity<BarnehagelisteResponse>
+    ): ResponseEntity<KindergartenlistResponse>
 
     @Operation(summary = "Get status for submitted kindergartnen list")
     @ApiResponses(
@@ -100,7 +100,7 @@ interface BarnehagelisterController {
                 content = [
                     Content(
                         mediaType = "application/json",
-                        schema = Schema(implementation = BarnehagelisteResponse::class),
+                        schema = Schema(implementation = KindergartenlistResponse::class),
                     ),
                 ],
             ),
@@ -110,7 +110,7 @@ interface BarnehagelisterController {
                 content = [
                     Content(
                         mediaType = "application/json",
-                        schema = Schema(implementation = BarnehagelisteResponse::class),
+                        schema = Schema(implementation = KindergartenlistResponse::class),
                     ),
                 ],
             ),
@@ -153,7 +153,7 @@ interface BarnehagelisterController {
     fun status(
         @PathVariable transaksjonsId: UUID,
         request: HttpServletRequest,
-    ): ResponseEntity<BarnehagelisteResponse>
+    ): ResponseEntity<KindergartenlistResponse>
 
     @GetMapping(
         path = ["/ping"],
@@ -161,90 +161,3 @@ interface BarnehagelisterController {
     )
     fun ping(): String
 }
-
-@Schema(name = "KindergartenListResponse")
-data class BarnehagelisteResponse(
-    val id: UUID,
-    val status: BarnehagelisteStatusEngelsk,
-    @Schema(name = "ReceivedTime")
-    val mottattTid: LocalDateTime,
-    @Schema(name = "FinishedTime")
-    val ferdigTid: LocalDateTime?,
-    val links: ResponseLinker,
-)
-
-enum class BarnehagelisteStatus(
-    val engelsk: BarnehagelisteStatusEngelsk,
-) {
-    MOTTATT(BarnehagelisteStatusEngelsk.RECEIVED),
-    FERDIG(BarnehagelisteStatusEngelsk.DONE),
-}
-
-enum class BarnehagelisteStatusEngelsk {
-    RECEIVED,
-    DONE,
-}
-
-@Schema(name = "ResponseLinks")
-data class ResponseLinker(
-    val status: String,
-)
-
-class ValideringsfeilException(
-    val errors: List<ValideringsfeilInfo>,
-) : RuntimeException("Validation error")
-
-@Schema(
-    name = "ValidationErrorInformation",
-    description = "Information about validation errors",
-)
-data class ValideringsfeilInfo(
-    @Schema(
-        description = "Which field has validation errors. If the info is unknown, then the field is set to missing",
-        example = "barnehager[0].navn",
-    )
-    val parameter: String,
-    @Schema(
-        description = "Details about the validation error. If no more information is available, the field is set to missing",
-        example = "must not be blank",
-    )
-    val detail: String,
-)
-
-@Schema(
-    name = "ProblemDetailWithCallIdAndErrors",
-    description = "Problem Details with callId and errors. Based on RFC 9457",
-)
-class ProblemDetailMedCallIdOgErrors(
-    @Schema(
-        description = "HTTP statuscode",
-        example = "400",
-    )
-    val status: Int,
-    @Schema(
-        description = "A short description of the error",
-        example = "Bad request",
-    )
-    val title: String,
-    @Schema(
-        description = "A readable description of the error",
-        example = "field must not be null",
-    )
-    val detail: String,
-    @Schema(
-        description = "A URI reference to the endpoint where the error occurred",
-        example = "/api/barnehagelister/v1",
-    )
-    val instance: String,
-    @Schema(
-        description = "A URI reference to a specific error type described at https://problems-registry.smartbear.com/",
-        example = "https://problems-registry.smartbear.com/validation-error/",
-    )
-    val type: String,
-    @Schema(
-        description = "An identifier for the error that can be used to track the error in later inquiries",
-        example = "57cf57cf06d84cc5883fc0a0a8804a7f",
-    )
-    val callId: String,
-    val errors: List<ValideringsfeilInfo>?,
-)
