+++ src/test/kotlin/no/nav/familie/ks/barnehagelister/rest/UnprotectedBarnehagelisteController.kt
@@ -3,6 +3,7 @@ package no.nav.familie.ks.barnehagelister.rest
 import jakarta.servlet.http.HttpServletRequest
 import no.nav.familie.ks.barnehagelister.domene.mapTilSkjemaV1
 import no.nav.familie.ks.barnehagelister.kontrakt.FormV1
+import no.nav.familie.ks.barnehagelister.kontrakt.KindergartenlistResponse
 import no.nav.security.token.support.core.api.Unprotected
 import org.springframework.context.annotation.Profile
 import org.springframework.http.ResponseEntity
@@ -20,13 +21,13 @@ class UnprotectedBarnehagelisteController(
         formV1: FormV1,
         bindingResult: BindingResult,
         request: HttpServletRequest,
-    ): ResponseEntity<BarnehagelisteResponse> = barnehagelisteService.mottaBarnehagelister(formV1.mapTilSkjemaV1(), bindingResult)
+    ): ResponseEntity<KindergartenlistResponse> = barnehagelisteService.mottaBarnehagelister(formV1.mapTilSkjemaV1(), bindingResult)
 
     @Unprotected
     override fun status(
         transaksjonsId: UUID,
         request: HttpServletRequest,
-    ): ResponseEntity<BarnehagelisteResponse> = barnehagelisteService.status(transaksjonsId)
+    ): ResponseEntity<KindergartenlistResponse> = barnehagelisteService.status(transaksjonsId)
 
     @Unprotected
     override fun ping(): String = barnehagelisteService.ping()
